boulder: Update to v0.25.6+git.7e13f93
This version enables the packager to use any valid git ref (such as HEAD,
a tag, a short git rev etc.), which will then be converted upon build
(emitting a warning when conversion happens) and saved in the recipe as
the canonical git revision hash.

Thank you to @barne856 for the feature development.

Signed-off-by: Rune Morling <ermo@serpentos.com>